refator: 댓글 신고 기능 comment 패키지로 이전
