--- src/main/java/dandi/dandi/commentreport/application/port/in/CommentReportUseCase.java
@@ -1,6 +1,0 @@
-package dandi.dandi.commentreport.application.port.in;
-
-public interface CommentReportUseCase {
-
-    void reportComment(Long memberId, Long commentId);
-}
